Set response cookie if a valid backend in the query string is present
Currently all cookies are session based (they expire when the browser is
closed). We currently use local storage to keep the information sent
from the custom headers which does not have expiration time.

Use case:
1. User visits site A - we set a cookie and local storage (both have value
X)
2. User closes the browser - this clears the cookie
3. User vists site A - we see there's X in local storage. So we use
this.
4. User visits site B - no cookie, no local storage, so we set the
cookie and local storage (both have value Y)
5. User now making requests on site A will take him from backend X to
backend Y because now a cookie is present which breaks the stickyness.

This PR makes sure we send a Set-Cookie header with the backend from the
query if it was valid.